Feat(Added Getmapping and DeleteMapping functionalities)

+++ src/main/java/com/greenfoxacademy/opal/kalendaryo/kalendaryo/controllers/AccountController.java
@@ -1,5 +1,6 @@
 package com.greenfoxacademy.opal.kalendaryo.kalendaryo.controllers;
 
+import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.api.UserResponse;
 import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.GoogleAuth;
 import com.greenfoxacademy.opal.kalendaryo.kalendaryo.model.entity.KalUser;
 import com.greenfoxacademy.opal.kalendaryo.kalendaryo.repository.GoogleAuthRepository;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
+import java.util.List;
 
 @RestController
 public class AccountController {
@@ -21,16 +23,26 @@ public class AccountController {
   AuthAndUserService authAndUserService;
 
   @GetMapping(value = "/account")
-  public ResponseEntity getGoogleAccounts (@RequestHeader("X-Client-Token") String clientToken, HttpServletRequest request) {
-
-    return new ResponseEntity(HttpStatus.OK);
+  public ResponseEntity getGoogleAccounts(@RequestHeader("X-Client-Token") String clientToken, HttpServletRequest request) {
+    if (!request.getHeader("X-Client-Token").equals("")) {
+      KalUser kalUser = authAndUserService.findUserByClientToken(clientToken);
+      List<GoogleAuth> googleAuthList;
+      googleAuthList = authAndUserService.listAllGoogleAccounts(kalUser.getId());
+      return ResponseEntity.status(200).body(new UserResponse(kalUser.getId(), kalUser.getUserEmail(), googleAuthList));
+    }
+    return ResponseEntity.status(401).body("Client token is missing or invalid");
   }
 
   @DeleteMapping(value = "/account")
-  public ResponseEntity deleteGoogleAccount (@RequestHeader("X-Client-Token") String clientToken, @RequestBody GoogleAuth googleAuth) throws IOException {
-    KalUser kalUser;
-    kalUser = authAndUserService.findUserByClientToken(clientToken);
-    authAndUserService.deleteGoogleAuth(googleAuth.getEmail(), kalUser);
-    return new ResponseEntity(HttpStatus.OK);
+  public ResponseEntity deleteGoogleAccount(@RequestHeader("X-Client-Token") String clientToken, @RequestBody GoogleAuth googleAuth, HttpServletRequest request) throws IOException {
+    if (!request.getHeader("X-Client-Token").equals("")) {
+      KalUser kalUser = authAndUserService.findUserByClientToken(clientToken);
+      if (!kalUser.getUserEmail().equals(googleAuth.getEmail())) {
+        authAndUserService.deleteGoogleAuth(googleAuth.getEmail(), kalUser);
+        return new ResponseEntity(HttpStatus.OK);
+      }
+      return ResponseEntity.status(500).body("Can not delete the base login email");
+    }
+    return ResponseEntity.status(401).body("Client token is missing or invalid");
   }
 }
